+++ src/control/mobile_manipulation/include/mobile_manipulation_rl/worlds.h
@@ -13,7 +13,7 @@
 #include <Eigen/Geometry>
 #include "tf/transform_datatypes.h"
 
-#include <modulation_rl/utils.h>
+#include <mobile_manipulation_rl/utils.h>
 
 class BaseWorld {
 public:
