+++ src/control/mobile_manipulation/src/ellipse.cpp
@@ -1,4 +1,4 @@
-#include "modulation_rl/ellipse.h"
+#include "mobile_manipulation_rl/ellipse.h"
 
 namespace ellipse {
