+++ src/control/mobile_manipulation/src/modulation_ellipses.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/modulation_ellipses.h>
+#include <mobile_manipulation_rl/modulation_ellipses.h>
 
 using namespace std;
 namespace modulation_ellipses {
