+++ src/control/mobile_manipulation/src/robot_hsr.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/robot_hsr.h>
+#include <mobile_manipulation_rl/robot_hsr.h>
 using Eigen::Affine3d;
 using Eigen::AngleAxisd;
 using Eigen::Translation3d;
