+++ src/control/mobile_manipulation/src/robot_pr2.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/robot_pr2.h>
+#include <mobile_manipulation_rl/robot_pr2.h>
 
 RobotPR2::RobotPR2(uint32_t seed, std::string strategy, std::string world_type, bool init_controllers, double penalty_scaling,
                    double time_step, bool perform_collision_check, std::string node_handle_name, bool verbose,
