+++ src/control/mobile_manipulation/src/robot_tiago.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/robot_tiago.h>
+#include <mobile_manipulation_rl/robot_tiago.h>
 
 RobotTiago::RobotTiago(uint32_t seed, std::string strategy, std::string world_type, bool init_controllers,
                        double penalty_scaling, double time_step, bool perform_collision_check, std::string node_handle_name,
