+++ src/control/mobile_manipulation/src/utils.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/utils.h>
+#include <mobile_manipulation_rl/utils.h>
 
 namespace utils {
   tf::Vector3 qToRpy(tf::Quaternion q) {
