+++ src/control/mobile_manipulation/src/worlds.cpp
@@ -1,4 +1,4 @@
-#include <modulation_rl/worlds.h>
+#include <mobile_manipulation_rl/worlds.h>
 
 BaseWorld::BaseWorld(std::string name, bool is_analytical) : name_{name}, is_analytical_{is_analytical} {
   if (name_ != "sim") {
